+++ test/svs/state_vector_test.go
@@ -70,7 +70,7 @@ func TestStateVectorEncodeDecode(t *testing.T) {
 
 func TestStateVectorDecodeStatic(t *testing.T) {
 	// WARNING: this does differ due to encoding entry's sources as strings and not as names, STC
-	comp, _ := enc.ComponentFromBytes([]byte{201, 16, 202, 3, 111, 110, 101, 204, 1, 1, 202, 3, 116, 119, 111, 204, 1, 2})
+	comp, _ := enc.ComponentFromBytes([]byte{201, 16, 202, 3, 111, 110, 101, 203, 1, 1, 202, 3, 116, 119, 111, 203, 1, 2})
 	sv, _ := svs.ParseStateVector(*comp)
 	assert.Equal(t, uint(1), sv.Get("one"))
 	assert.Equal(t, uint(2), sv.Get("two"))