+++ docs/CHANGELOG.md
@@ -6,9 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Go's Versioning](https://go.dev/doc/modules/release-workflow).
 
 ## [Unreleased]
+
+## [v0.0.0-alpha.3] - 2022-10-26
 ### Changed
-- Pulled out init() into its own file.
-- Utilize TLNum (instead of uint) for TlvTypes.
+- SVS: Pulled out init() into its own file.
+- Utilize TLNum (instead of uint) for SVS TlvTypes.
 - Fixed StateVector Encoding to met specification.
 
 ## [v0.0.0-alpha.2] - 2022-10-22
@@ -30,6 +32,7 @@ and this project adheres to [Go's Versioning](https://go.dev/doc/modules/release
 ### Security
 - SVS does is not secure due to having lack signing / validating capabilities (waiting on go-ndn)
 
-[Unreleased]: https://github.com/justincpresley/ndn-sync/compare/v0.0.0-alpha.2...HEAD
+[Unreleased]: https://github.com/justincpresley/ndn-sync/compare/v0.0.0-alpha.3...HEAD
+[v0.0.0-alpha.3]: https://github.com/justincpresley/ndn-sync/compare/v0.0.0-alpha.2...v0.0.0-alpha.3
 [v0.0.0-alpha.2]: https://github.com/justincpresley/ndn-sync/compare/v0.0.0-alpha.1...v0.0.0-alpha.2
 [v0.0.0-alpha.1]: https://github.com/justincpresley/ndn-sync/releases/tag/v0.0.0-alpha.1